requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
chromadb
PyMuPDF
python-docx
python-pptx
ollama
